update project section
