+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "framer-motion": "^12.6.2",
+    "gsap": "^3.12.7",
     "next": "15.2.4",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
