+++ src/app/globals.css
@@ -15,4 +15,5 @@ body {
   background: var(--background);
   color: var(--foreground);
   font-family: "Fira Code", monospace;
+  margin: 0;
 }
